Implemented graph auto updating (#109)
This one was an absolute beast to implement and took me almost 6 hours straight of researching, testing, and implementing different strategies until I finally hit the sweet one liner that worked. Graph Y axis now updates upon cubes being drawn and this value is drawn out for 10 frames. This could easily be increased to 100 and no user should ever notice that its coded in a sense that you are "limited" on frames. Also when cubes are changed in frames greater than 1 the graph updates on the fly with the reductions and increases of energies.
References #104